Ensure that test passwords confirm to the strong rules
